Rollback to working state
